Estimate conflict scores for n subjects

Estimate conflict scores for n subjects

c_scores(
  subjects,
  c_nmst = c(n, mu, sigma, tau),
  nc_nmst = c(n, mu, sigma, tau)
)

Arguments

subjects
A number for the number conflict scores to generate

c_nmst
A vector containing the parameters for from an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.

nc_nmst
A vector containing the parameters for sampling from an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.

Value

A simulated conflict score for one subject

Details

This function creates simulated conflict scores for n number of subjects. This is a wrapper for c_score, which creates a single conflict score.

Examples

c_scores(subjects=50,
        c_nmst= c(n=10, mu=600, sigma=100, tau=100),
        nc_nmst= c(n=10, mu=500, sigma=100, tau=100))
#>  [1] 191.262500  99.439406  58.510046  42.908417  68.627690  97.131166
#>  [7] 100.769236 161.056118  54.973075  37.548866 116.175906  48.476117
#> [13]  -5.987813 154.133365 161.458789 163.436768 158.477082 133.146871
#> [19] 136.337707 102.139070  98.853074  76.745309 286.191525 213.347679
#> [25] 135.807677 148.406541  78.592845 167.551427  27.096289  96.499994
#> [31]  56.726197  68.825956  59.961426  64.724301 128.757382 139.606625
#> [37]  96.175124  95.282824  67.527541  67.257763  97.709475 202.178031
#> [43]  55.618978  78.376055 252.425101  70.456493   3.384013  -8.195291
#> [49] 189.005670 193.449662