Estimate conflict scores for n subjects

Estimate conflict scores for n subjects

c_scores(subjects, c_nmst = c(n, mu, sigma, tau), nc_nmst = c(n, mu,
  sigma, tau))

Arguments

subjects
A number for the number conflict scores to generate

c_nmst
A vector containing the parameters for from an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.

nc_nmst
A vector containing the parameters for sampling from an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.

Value

A simulated conflict score for one subject

Details

This function creates simulated conflict scores for n number of subjects. This is a wrapper for c_score, which creates a single conflict score.

Examples

c_scores(subjects=50,
        c_nmst= c(n=10, mu=600, sigma=100, tau=100),
        nc_nmst= c(n=10, mu=500, sigma=100, tau=100))

#>  [1]  19.38388  46.98642 111.79709  63.35863 240.71755  22.57877  26.91940
#>  [8] 129.46284 168.95506  51.01656 205.48587  78.31243 127.59321 148.91130
#> [15]  25.19426 140.73002 113.19346 135.89940  92.92195  59.76814 102.40999
#> [22] 189.92489 104.02119  70.32033  84.21691  74.83226 153.76440  70.75507
#> [29] -33.73567 128.75658 127.89730 102.00309 160.39661  60.73642 100.05715
#> [36] 124.20202 206.24455 163.49637  94.74934 -37.26411  92.32906 226.29088
#> [43] 127.57589  95.34577  30.11422 135.67215 -15.81387 199.23142  30.06636
#> [50] 209.49678